Cleanup, boss, menus, aim and Enemy Scaling

+++ Scripts/Damageable.cs
@@ -4,14 +4,14 @@ using System;
 public partial class Damageable : CharacterBody3D
 {
     [Export] public Player player;
-    [Export] public int MaxHealth;
-    public int CurrentHealth { get; set; }
+    [Export] public double MaxHealth;
+    public double CurrentHealth { get; set; }
 
     public override void _Ready()
     {
     }
 
-    public virtual void HandleHit(int damage)
+    public virtual void HandleHit(double damage)
     {
         CurrentHealth -= damage;
         if (CurrentHealth <= 0)
